chore: improve Progress exports
